requirements.txt
llama-index==0.12.43
llama-index-embeddings-voyageai==0.3.6
llama_index_embeddings_huggingface==0.5.4
llama-index-vector-stores-pinecone==0.5.0
pinecone-client==6.0.0
python-dotenv==1.1.0
requests==2.32.4
streamlit==1.46.0
voyageai==0.3.2 